requirements.txt
pandas
firebase-admin
cloudinary
xlsxwriter
streamlit-cookies-manager
streamlit-authenticator
python-docx
streamlit==1.53.0
protobuf>=4.25.0



